Split system name tests since it's crashing the env
